requirements-azure-python.txt
adal==1.2.7
azure-common==1.1.27
azure-core==1.14.0
azure-identity==1.6.0
azure-mgmt-compute==10.0.0
azure-mgmt-core==1.2.2
azure-mgmt-dns==2.1.0
azure-mgmt-monitor==1.0.0
azure-mgmt-network==12.0.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-rdbms==1.9.0
azure-mgmt-resource==2.2.0
azure-nspkg==3.0.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.10
isodate==0.6.0
msal==1.12.0
msal-extensions==0.3.0
msrest==0.6.21
msrestazure==0.6.4
mysql-connector==2.2.9
oauthlib==3.1.0
portalocker==1.7.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python-decouple==3.4
requests==2.25.1
requests-oauthlib==1.3.0
semantic-version==2.8.5
setuptools-rust==0.12.1
six==1.16.0
toml==0.10.2
urllib3==1.26.5